+++ Source/Scene/Batched3DModel3DTileContent.js
@@ -495,10 +495,6 @@ define([
         if (defined(tilesetClippingPlanes)) {
             // Dereference the clipping planes from the model if they are irrelevant - saves on shading
             this._model.clippingPlanes = tilesetClippingPlanes.enabled && this._tile._isClipped ? tilesetClippingPlanes : undefined;
-        } else if (defined(this._model.clippingPlanes)) {
-            // Destroy any other clipping planes on the model if the tileset doesn't have any clipping planes
-            this._model.clippingPlanes.checkDestroy();
-            this._model.clippingPlanes = undefined;
         }
 
         this._model.update(frameState);