+++ Specs/Scene/Cesium3DTilesetSpec.js
@@ -2872,6 +2872,53 @@ defineSuite([
         });
     });
 
+    it('takes ownership of ClippingPlaneCollections', function() {
+        return Cesium3DTilesTester.loadTileset(scene, tilesetUrl).then(function(tileset) {
+            var clippingPlaneCollection1 = new ClippingPlaneCollection({
+                planes : [
+                    new Plane(Cartesian3.UNIT_Z, -100000000.0)
+                ]
+            });
+            expect(clippingPlaneCollection1.owner).not.toBeDefined();
+
+            tileset.clippingPlanes = clippingPlaneCollection1;
+            expect(clippingPlaneCollection1.owner).toBe(tileset);
+
+            var clippingPlaneCollection2 = new ClippingPlaneCollection({
+                planes : [
+                    new Plane(Cartesian3.UNIT_Z, -100000000.0)
+                ]
+            });
+
+            tileset.clippingPlanes = clippingPlaneCollection2;
+            expect(clippingPlaneCollection2.owner).toBe(tileset);
+        });
+    });
+
+    it('handles destroying ClippingPlaneCollections', function() {
+        return Cesium3DTilesTester.loadTileset(scene, tilesetUrl).then(function(tileset) {
+            var clippingPlaneCollection1 = new ClippingPlaneCollection({
+                planes : [
+                    new Plane(Cartesian3.UNIT_Z, -100000000.0)
+                ]
+            });
+            expect(clippingPlaneCollection1.owner).not.toBeDefined();
+
+            tileset.clippingPlanes = clippingPlaneCollection1;
+            var clippingPlaneCollection2 = new ClippingPlaneCollection({
+                planes : [
+                    new Plane(Cartesian3.UNIT_Z, -100000000.0)
+                ]
+            });
+
+            tileset.clippingPlanes = clippingPlaneCollection2;
+            expect(clippingPlaneCollection1.isDestroyed()).toBe(true);
+
+            scene.primitives.remove(tileset);
+            expect(clippingPlaneCollection2.isDestroyed()).toBe(true);
+        });
+    });
+
     it('clipping planes cull hidden tiles', function() {
         return Cesium3DTilesTester.loadTileset(scene, tilesetUrl).then(function(tileset) {
             var visibility = tileset._root.visibility(scene.frameState, CullingVolume.MASK_INSIDE);